Added new feature: tests carried out with test.py will now be identified with specific names
When using `bosh test`, pytest will now assign an id for each of the generated tests. For a single test, this id is: name of the descriptor concatenated with the name of the test case.

Among other things, through the pytest output, users will be able to identify what are the defined test cases that failed.