Fixing issue #423 by removing arguments.cwl and better showing intent.
 - removed `tests/cwl/stdout/arguments.cwl`
 - renamed `tests/cwl/array-output/` to `tests/cwl/array-outputs/` for
consistency
 - renamed all `.yml` files under `tests/cwl/` to the same base name as
their containing directory (mostly moving the `-job1` stuff)
 - improved the visible intent of the `test_import_cwl_valid` method in
`test_import.py` for collecting `.cwl` and `.yml` files. The `.cwl` file
is not loaded based on the containing directory name and if a `.yml'
file exists within the containing directory, it gets loaded too.
 - Added some description of the file structure in `tests/cwl/README.md`
to improve intent and give an idea how it should be used. Also mentioned
bad_files and that this lists the folders under `tests/cwl/` that we
expect to throw exceptions. _I don't know why it is this way, but now
someone as clueless as me can ask the right questions._