Fixed issue where some singularity containers suspended pything
This closes #431

 - Adds `-contain` to singularity opts if it or `-containall` aren't
already there
 - sets `conOpts` to `[]` if it is None for simpler control-flow later

 I am not sure if boutiques should be allowing singularity to run
without `-contains` but there is plausibly a reason, and if it comes up
we might need a means to allow this, maybe with a flag? As this seems to
only affect certain singularity containers (that pollute the parent
processes signals, it seems), there is an argument to be made that the
descriptor author should be responsible... it's just such an esoteric
and hard-to-catch issue that I think defaulting to this until someone
asks makes more sense, as someone asking deliberately intends to
communicate signals to the parent process, so they'll be able to ask the
boutques team the correct questions.